requirements.txt
pandas
sqlalchemy
psycopg2-binary
requests
openpyxl
python-dotenv
pyarrow
click
pytest
azure-identity
azure-storage-file-datalake
